requirements.txt
bcrypt==4.1.1
blinker==1.7.0
click==8.1.7
debugpy==1.8.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
